requirements.txt


peft==0.15.1
POT==0.9.5
PyYAML==6.0.2
PyYAML==6.0.2
scikit_learn==1.7.0
scipy==1.15.3
seaborn==0.13.2
torch==2.6.0
tqdm==4.67.1
transformers==4.51.0
geomloss==0.2.6
matplotlib==3.10.3